refactor: split CLI into package with core-only dependencies
