+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
 dev = ["pytest>=7.0", "ruff>=0.4", "mypy>=1.10", "setuptools-scm>=8", "pyinstaller>=6.0", "pytest-timeout>=2.0"]
 
 [project.scripts]
-cogstash = "cogstash:main"
+cogstash = "cogstash.cli.__main__:main"
 
 [tool.setuptools.packages.find]
 where = ["src"]
