+++ src/cogstash/cli/windows.py
@@ -1,0 +1,3 @@
+from cogstash._windows import prepare_windows_cli_console
+
+__all__ = ["prepare_windows_cli_console"]
